+++ ML-Frameworks/pytorch-aarch64/build.sh
@@ -11,15 +11,11 @@ build_log=build-$(git rev-parse --short=7 HEAD)-$(date '+%Y-%m-%dT%H-%M-%S').log
 exec &> >(tee -a "$build_log")
 
 # Bail out if sources are already there
-if [ -f .torch_build_container_id ] || [ -f .torch_ao_build_container_id ] || \
-    [ -d ao ] || [ -d ComputeLibrary ] || [ -d pytorch ]; then
-    printf "\n\n%s\n%s\n%s\n%s\n%s\n%s\n%s\n\n\n" \
+if [ -f .torch_build_container_id ] || [ -d pytorch ]; then
+    printf "\n\n%s\n%s\n%s\n%s\n\n\n" \
         "You appear to have artefacts from a previous build lying around." \
         "Check for any of the following:" \
         "  - .torch_build_container_id" \
-        "  - .torch_ao_build_container_id" \
-        "  - ao" \
-        "  - ComputeLibrary" \
         "  - pytorch"
 
     if [[ "$*" != *--fresh* ]] && [[ "$*" != *--use-existing-sources* ]]; then
@@ -38,7 +34,6 @@ if [ -f .torch_build_container_id ] || [ -f .torch_ao_build_container_id ] || \
             if [ ! -z "$(docker ps -a --no-trunc | grep $TORCH_BUILD_CONTAINER)" ]; then
                 # Change permissions from root
                 docker exec $TORCH_BUILD_CONTAINER chown -R $(id -u):$(id -g) /artifacts 2>/dev/null || true
-                docker exec $TORCH_BUILD_CONTAINER chown -R $(id -u):$(id -g) /ComputeLibrary 2>/dev/null || true
                 docker exec $TORCH_BUILD_CONTAINER chown -R $(id -u):$(id -g) /pytorch 2>/dev/null || true
 
                 # Wipe old container
@@ -47,21 +42,7 @@ if [ -f .torch_build_container_id ] || [ -f .torch_ao_build_container_id ] || \
             rm -f .torch_build_container_id
         fi
 
-        # Make sure we can wipe directories created with root privileges in Docker
-        if [ -f .torch_ao_build_container_id ]; then
-            TORCH_AO_BUILD_CONTAINER=$(cat .torch_ao_build_container_id)
-            if [ ! -z "$(docker ps -a --no-trunc | grep $TORCH_AO_BUILD_CONTAINER)" ]; then
-                docker exec $TORCH_AO_BUILD_CONTAINER chown -R $(id -u):$(id -g) /ao
-
-                # Wipe old container
-                docker rm -f $TORCH_AO_BUILD_CONTAINER 2>/dev/null || true
-            fi
-            rm -f .torch_ao_build_container_id
-        fi
-
         # Wipe the other directories; we should have the privileges now
-        if [ -d ao ]; then rm -rf ao; fi
-        if [ -d ComputeLibrary ]; then rm -rf ComputeLibrary; fi
         if [ -d pytorch ]; then rm -rf pytorch; fi
     fi
 fi
@@ -86,9 +67,4 @@ fi
 # Use the second to last match, otherwise grep finds itself
 torch_wheel_name=$(grep -o "torch-.*.whl" "$build_log" | head -n -1 | tail -n 1)
 
-./build-torch-ao-wheel.sh
-
-# Use the second to last match, otherwise grep finds itself
-torch_ao_wheel_name=$(grep -o "torchao-.*.whl" "$build_log" | head -n -1 | tail -n 1)
-
-./dockerize.sh "results/$torch_wheel_name" "results/$torch_ao_wheel_name" --build-only
+./dockerize.sh "results/${torch_wheel_name}" --build-only